# create a worktree-tree without an intermediate index