--- a/crates/gitbutler-repo/Cargo.toml
+++ b/crates/gitbutler-repo/Cargo.toml
@@ -7,7 +7,7 @@ publish = false
 
 [dependencies]
 git2.workspace = true
-gix.workspace = true
+gix = { workspace = true, features = ["status", "tree-editor"] }
 anyhow = "1.0.86"
 bstr.workspace = true
 tokio = { workspace = true, features = [